Add files via upload
This is my second project on Numpy. 
Project title is "Analyse London Olympics Dataset".
DESCRIPTION

Problem: 

Evaluate the dataset of the Summer Olympics, London 2012 to:

Find and print the name of the country that won maximum gold medals,
Find and print the countries who won more than 20 gold medals,
Print the medal tally,
Print each country name with the corresponding number of gold medals, and
Print each country's name with the total number of medals won.