Add files via upload
Problem:

The given dataset lists the glucose level readings of several pregnant women taken either during a survey examination or routine medical care. It specifies if the 2-hour post-load plasma glucose was at least 200 mg/dl. Analyze the dataset to:

Find the features of the dataset,
Find the response label of the dataset,
Create a model  to predict the diabetes outcome,
Use training and testing datasets to train the model, and
Check the accuracy of the model.